+++ client/package.json
@@ -24,6 +24,7 @@
     "react-redux": "^8.0.2",
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
+    "react-table": "^7.8.0",
     "react-toastify": "^9.0.5",
     "toastify": "^2.0.1",
     "web-vitals": "^2.1.0"
