+++ client/src/components/table/table.css
@@ -1,0 +1,23 @@
+table {
+  font-family: Arial, Helvetica, sans-serif;
+  border-collapse: collapse;
+  width: 100%;
+  height: 100%;
+}
+
+table td, table th {
+  border: 1px solid #ddd;
+  padding: 0.5em;
+}
+
+table tr:nth-child(even){background-color: #f2f2f2;}
+
+table tr:hover {background-color: #ddd;}
+
+table th, tfoot {
+  padding-top: 1em;
+  padding-bottom: 1em;
+  text-align: center;
+  background-color:darkcyan;
+  color: white;
+}
