--- client/src/data2.js
@@ -1,47 +1,0 @@
-export const MyData = [
-  {
-    id: 1,
-    x: 3,
-    y: 4,
-  },
-  {
-    id: 2,
-    x: 2,
-    y: 3,
-  },
-  {
-    id: 3,
-    x: 5,
-    y: 2,
-  },
-  {
-    id: 4,
-    x: 4,
-    y: 1,
-  },
-  {
-    id: 5,
-    x: 3,
-    y: 6,
-  },
-  {
-    id: 6,
-    x: 1,
-    y: 3,
-  },
-  {
-    id: 7,
-    x: 2,
-    y: 1,
-  },
-  {
-    id: 8,
-    x: 3,
-    y: 6,
-  },
-  {
-    id: 9,
-    x: 7,
-    y: 9,
-  }
-];
