+++ client/src/index.js
@@ -3,14 +3,14 @@ import ReactDOM from 'react-dom/client';
 import '@fortawesome/fontawesome-free/js/all.js';
 import './index.css';
 import App from './app';
-//import {Provider} from 'react-redux';
-
+import {Provider} from 'react-redux';
+import store from './store';
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
-
+    <Provider store={store}>
       <App />
-
+    </Provider>
   </React.StrictMode>
 );
