+++ client/src/optionsSlice.js
@@ -1,0 +1,18 @@
+import { createSlice } from '@reduxjs/toolkit';
+
+const initialState = {
+  options: ["conditionals", "loops"]
+}
+
+const optionsSlice = createSlice({
+  name: "options",
+  initialState,
+  reducers: {
+    change: (state, action) => {
+      state.options = action.payload;
+    }
+  }
+})
+
+export const {change} = optionsSlice.actions;
+export default optionsSlice.reducer;
