+++ server/public/output.json
@@ -1,0 +1,79 @@
+{
+    "constants.py": {
+        "conditionals": 0,
+        "loops": 0,
+        "loops_for": 0,
+        "loops_while": 0,
+        "functions": 0,
+        "functions_recursion": 0,
+        "max_depth": 0,
+        "avg_depth": 0,
+        "sum_depth": 0
+    },
+    "elegance.py": {
+        "conditionals": 1,
+        "loops": 6,
+        "loops_for": 6,
+        "loops_while": 0,
+        "functions": 8,
+        "functions_recursion": 0,
+        "max_depth": 3,
+        "avg_depth": 2.11,
+        "sum_depth": 19
+    },
+    "main.py": {
+        "conditionals": 7,
+        "loops": 9,
+        "loops_for": 9,
+        "loops_while": 0,
+        "functions": 1,
+        "functions_recursion": 1,
+        "max_depth": 16,
+        "avg_depth": 16.0,
+        "sum_depth": 16
+    },
+    "output.json": {
+        "conditionals": 0,
+        "loops": 0,
+        "loops_for": 0,
+        "loops_while": 0,
+        "functions": 0,
+        "functions_recursion": 0,
+        "max_depth": 0,
+        "avg_depth": 0,
+        "sum_depth": 0
+    },
+    "PythonCodeAnalyzer.py": {
+        "conditionals": 13,
+        "loops": 12,
+        "loops_for": 9,
+        "loops_while": 3,
+        "functions": 5,
+        "functions_recursion": 0,
+        "max_depth": 8,
+        "avg_depth": 6.67,
+        "sum_depth": 40
+    },
+    "PythonLexer.py": {
+        "conditionals": 8,
+        "loops": 4,
+        "loops_for": 2,
+        "loops_while": 2,
+        "functions": 12,
+        "functions_recursion": 0,
+        "max_depth": 14,
+        "avg_depth": 2.69,
+        "sum_depth": 35
+    },
+    "utils.py": {
+        "conditionals": 1,
+        "loops": 3,
+        "loops_for": 3,
+        "loops_while": 0,
+        "functions": 3,
+        "functions_recursion": 1,
+        "max_depth": 4,
+        "avg_depth": 3.0,
+        "sum_depth": 9
+    }
+}
